Validate file and directory access based on file permissions

+++ src/main/groovy/org/mockftpserver/fake/command/DeleCommandHandler.groovy
@@ -44,10 +44,7 @@ class DeleCommandHandler extends AbstractFakeCommandHandler {
         verifyFileSystemCondition(fileSystem.isFile(path), path, 'filesystem.isNotAFile')
 
         // User must have write permission to the parent directory
-        def userAccount = getUserAccount(session)
-        def parent = fileSystem.getParent(path)
-        def entry = fileSystem.getEntry(parent)
-        verifyFileSystemCondition(userAccount.canWrite(entry), parent, 'filesystem.cannotWrite')
+        verifyWritePermission(session, fileSystem.getParent(path))
 
         fileSystem.delete(path)
         sendReply(session, ReplyCodes.DELE_OK, 'dele', [path])
